chore: format solidity

+++ contracts/src/ParametricToken.sol
@@ -62,8 +62,8 @@ contract ParametricToken is ERC20, IParametricToken {
     }
 
     constructor(string memory _name, string memory _symbol) ERC20(_name, _symbol) {
-        PARAM_CONFIG = [ParamConfig({ name: bytes32("myParam"), decimals: 0, isMutable: true })];
-        for (uint i = 0; i < NUMBER_OF_PARAMETERS; i++) {
+        PARAM_CONFIG = [ParamConfig({name: bytes32("myParam"), decimals: 0, isMutable: true})];
+        for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
             _parametersInit[i] = 0;
         }
     }
@@ -168,7 +168,7 @@ contract ParametricToken is ERC20, IParametricToken {
         acc.accountType = AccountType.Super;
 
         // Create subId 0 with current balance
-        _supers[account].subs.push(SubAccount({ balance: acc.balance, parameters: acc.parameters }));
+        _supers[account].subs.push(SubAccount({balance: acc.balance, parameters: acc.parameters}));
         _supers[account].subsCount = 1;
 
         // Clear normal parameters
@@ -185,7 +185,7 @@ contract ParametricToken is ERC20, IParametricToken {
 
         SuperAccount storage acc = _supers[account];
 
-        acc.subs.push(SubAccount({ balance: 0, parameters: _parametersInit }));
+        acc.subs.push(SubAccount({balance: 0, parameters: _parametersInit}));
         acc.subsCount = uint48(acc.subs.length);
         uint48 newSubId = acc.subsCount - 1;
 
@@ -214,14 +214,24 @@ contract ParametricToken is ERC20, IParametricToken {
         return _accounts[account].parameters[paramIndex];
     }
 
-    function parameterOfSub(uint8 paramIndex, address account, uint48 subId) external view onlyValidSub(account, subId) returns (uint64) {
+    function parameterOfSub(uint8 paramIndex, address account, uint48 subId)
+        external
+        view
+        onlyValidSub(account, subId)
+        returns (uint64)
+    {
         require(paramIndex < NUMBER_OF_PARAMETERS, "Index exceeds number of parameters");
         return _supers[account].subs[subId].parameters[paramIndex];
     }
 
     // ========== Allowances ==========
 
-    function allowanceForSub(address owner, uint48 subId, address spender) external view onlyValidSub(owner, subId) returns (uint256) {
+    function allowanceForSub(address owner, uint48 subId, address spender)
+        external
+        view
+        onlyValidSub(owner, subId)
+        returns (uint256)
+    {
         Allowance storage al = _allowances[owner][spender];
         if (al.subId == subId) {
             return al.sub;
@@ -251,17 +261,20 @@ contract ParametricToken is ERC20, IParametricToken {
     }
 
     // Helper to check and consume allowance for a specific subId
-    function _sufficientAllowanceForSub(
-        address owner,
-        address spender,
-        uint48 fromSubId,
-        uint256 amount
-    ) internal view onlyValidSub(owner, fromSubId) returns (bool) {
+    function _sufficientAllowanceForSub(address owner, address spender, uint48 fromSubId, uint256 amount)
+        internal
+        view
+        onlyValidSub(owner, fromSubId)
+        returns (bool)
+    {
         Allowance storage al = _allowances[owner][spender];
         return fromSubId == al.subId ? al.sub >= amount : al.total - al.sub >= amount;
     }
 
-    function _consumeAllowanceForSub(address owner, address spender, uint48 fromSubId, uint256 amount) internal onlyValidSub(owner, fromSubId) {
+    function _consumeAllowanceForSub(address owner, address spender, uint48 fromSubId, uint256 amount)
+        internal
+        onlyValidSub(owner, fromSubId)
+    {
         Allowance storage al = _allowances[owner][spender];
 
         if (fromSubId == al.subId) {
@@ -304,7 +317,11 @@ contract ParametricToken is ERC20, IParametricToken {
         return true;
     }
 
-    function _weightedAverage(uint64 param1, uint256 amount1, uint64 param2, uint256 amount2) private pure returns (uint64) {
+    function _weightedAverage(uint64 param1, uint256 amount1, uint64 param2, uint256 amount2)
+        private
+        pure
+        returns (uint64)
+    {
         require(amount1 > 0 && amount2 > 0, "Invalid amounts");
         uint256 sumProduct = uint256(param1) * amount1 + uint256(param2) * amount2;
         uint256 sum = amount1 + amount2;
@@ -314,7 +331,11 @@ contract ParametricToken is ERC20, IParametricToken {
 
     // ========== Transfers ==========
 
-    function transferToSub(address toSuper, uint48 toSubId, uint256 amount) external onlyValidSub(toSuper, toSubId) returns (bool) {
+    function transferToSub(address toSuper, uint48 toSubId, uint256 amount)
+        external
+        onlyValidSub(toSuper, toSubId)
+        returns (bool)
+    {
         address from = _msgSender();
         Account storage fromAcc = _accounts[from];
         Account storage toAcc = _accounts[toSuper];
@@ -335,7 +356,8 @@ contract ParametricToken is ERC20, IParametricToken {
         if (oldSubBalance > 0) {
             for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
                 if (PARAM_CONFIG[i].isMutable) {
-                    toSubAcc.parameters[i] = _weightedAverage(toSubAcc.parameters[i], oldSubBalance, fromAcc.parameters[i], amount);
+                    toSubAcc.parameters[i] =
+                        _weightedAverage(toSubAcc.parameters[i], oldSubBalance, fromAcc.parameters[i], amount);
                 }
             }
         } else {
@@ -374,7 +396,9 @@ contract ParametricToken is ERC20, IParametricToken {
         if (oldToBalance > 0) {
             for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
                 if (PARAM_CONFIG[i].isMutable) {
-                    toAcc.parameters[i] = _weightedAverage(toAcc.parameters[i], oldToBalance, fromSuperAcc.subs[fromSubId].parameters[i], amount);
+                    toAcc.parameters[i] = _weightedAverage(
+                        toAcc.parameters[i], oldToBalance, fromSuperAcc.subs[fromSubId].parameters[i], amount
+                    );
                 }
             }
         } else {
@@ -435,7 +459,11 @@ contract ParametricToken is ERC20, IParametricToken {
 
     // ========== Approved Transfers ==========
 
-    function approvedTransferToSub(address from, address toSuper, uint48 toSubId, uint256 amount) external onlyValidSub(toSuper, toSubId) returns (bool) {
+    function approvedTransferToSub(address from, address toSuper, uint48 toSubId, uint256 amount)
+        external
+        onlyValidSub(toSuper, toSubId)
+        returns (bool)
+    {
         address spender = _msgSender();
 
         // Execute transfer
@@ -458,7 +486,8 @@ contract ParametricToken is ERC20, IParametricToken {
         if (oldSubBalance > 0) {
             for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
                 if (PARAM_CONFIG[i].isMutable) {
-                    toSubAcc.parameters[i] = _weightedAverage(toSubAcc.parameters[i], oldSubBalance, fromAcc.parameters[i], amount);
+                    toSubAcc.parameters[i] =
+                        _weightedAverage(toSubAcc.parameters[i], oldSubBalance, fromAcc.parameters[i], amount);
                 }
             }
         } else {
@@ -505,7 +534,9 @@ contract ParametricToken is ERC20, IParametricToken {
         if (oldSubBalance > 0) {
             for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
                 if (PARAM_CONFIG[i].isMutable) {
-                    toSubAcc.parameters[i] = _weightedAverage(toSubAcc.parameters[i], oldSubBalance, fromSuperAcc.subs[fromSubId].parameters[i], amount);
+                    toSubAcc.parameters[i] = _weightedAverage(
+                        toSubAcc.parameters[i], oldSubBalance, fromSuperAcc.subs[fromSubId].parameters[i], amount
+                    );
                 }
             }
         } else {
@@ -551,14 +582,15 @@ contract ParametricToken is ERC20, IParametricToken {
             // Calculate new weighted average parameters
             uint256 oldBalance = sub0.balance;
 
-            for (uint i = 0; i < NUMBER_OF_PARAMETERS; i++) {
+            for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
                 if (PARAM_CONFIG[i].isMutable) {
                     if (oldBalance == 0) {
                         // First mint to this sub - set to block.timestamp
                         sub0.parameters[i] = uint64(block.timestamp);
                     } else {
                         // Weighted average for non-zero
-                        sub0.parameters[i] = _weightedAverage(sub0.parameters[i], oldBalance, uint64(block.timestamp), amount);
+                        sub0.parameters[i] =
+                            _weightedAverage(sub0.parameters[i], oldBalance, uint64(block.timestamp), amount);
                     }
                 } else {
                     // Immutable parameter
@@ -579,14 +611,15 @@ contract ParametricToken is ERC20, IParametricToken {
             // Normal account
             uint256 oldBalance = acc.balance;
 
-            for (uint i = 0; i < NUMBER_OF_PARAMETERS; i++) {
+            for (uint256 i = 0; i < NUMBER_OF_PARAMETERS; i++) {
                 if (PARAM_CONFIG[i].isMutable) {
                     if (oldBalance == 0) {
                         // First mint - set to block.timestamp
                         acc.parameters[i] = uint64(block.timestamp);
                     } else {
                         // Non-zero balance
-                        acc.parameters[i] = _weightedAverage(acc.parameters[i], oldBalance, uint64(block.timestamp), amount);
+                        acc.parameters[i] =
+                            _weightedAverage(acc.parameters[i], oldBalance, uint64(block.timestamp), amount);
                     }
                 } else {
                     // Immutable parameter
